Move agOptions to local model
